Feat: <DMA接收>
使能DMA串口接收，每接收1000字节 产生一次中断，使用DMA可以减少中断次数，增加系统流畅度

+++ Code/HARDWARE/DMA/dma.h
@@ -5,7 +5,7 @@
 
 #define BUFFER_SIZE	32
 extern const u8 SRC_Buffer[];
-extern u32 DST_Buffer[];
+extern u8 DST_Buffer[];
 
 void Usart1_DMA2_init(void);
 void DMA2_init(void);
